requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
yfinance>=0.2.0
ta>=0.10.0
python-dotenv>=0.19.0
feedparser>=6.0.0
scipy>=1.9.0
